test/requirements.txt
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
iniconfig==1.1.1
isodate==0.6.0
jsonschema==3.2.0
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.0
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.4
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.4
